requirements.txt
Keras==2.12.0
numpy==1.23.5
Pillow==9.4.0
Flask==2.3.2